move manifest features to separate directories and use dynamic import to implement code splitting
