+++ client/src/features/manifestDetails/index.ts
@@ -1,0 +1,3 @@
+import { ManifestDetails as Component } from './ManifestDetails';
+
+export { Component };
