+++ client/src/features/manifestList/index.ts
@@ -1,0 +1,3 @@
+import { ManifestList as Component } from './ManifestList';
+
+export { Component };
