+++ client/src/features/newManifest/index.ts
@@ -1,0 +1,3 @@
+import { NewManifest as Component } from './NewManifest';
+
+export { Component };
